LikeWiseApp/requirements.txt
torch
torchvision
numpy
Pillow
pandas
flask
opencv-python